backend/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
python-dotenv==1.0.0
Pillow==10.1.0
django-rest-knox==4.2.0 